fix(storage): stream per-snapshot writes, close sniff TOCTOU window
Two more real findings from a fresh Codex review round on PR #869:

1. write_bundle_facts_archive()'s per-library-snapshot encode still
   called json.dumps()+.encode() directly, materializing a full copy of
   one oversized snapshot's serialization before the aggregate cap
   check (checked only after) ever got a chance to reject it. Fixed by
   routing it through a new bounded_encode_utf8() helper, which streams
   against the remaining aggregate allowance: a single oversized string
   field is caught via the existing oversized_raw_string() fast path,
   and the aggregate across many individually-bounded fields is caught
   by streaming JSONEncoder.iterencode() and stopping as soon as the
   running byte count crosses the limit.

2. sniff_bundle_archive_format() still did its own, separate
   Path.stat() then a separate open() -- the exact two-inode TOCTOU
   window open_regular_file_for_format_sniff()'s own identical race had
   already been fixed for, just never carried over to this sibling
   function. Fixed by delegating to that helper's single
   O_NONBLOCK-open-then-fstat() classification.

5 new regression tests, each confirmed to fail against the pre-fix
code before the fix was applied (including a structural test proving
sniff_bundle_archive_format() never calls a separate Path.stat(), and
a tracking-wrapper test proving the per-snapshot encode routes through
bounded_encode_utf8()).

tests/test_bundle_facts_archive_hardening.py (new): every resource-
limit/cap/duplicate-charging test split out of
tests/test_bundle_facts_archive.py, purely to keep both files under
the ADR-061 1200-line test cap after these additions.

Full combined test suite green (142 passed), mypy/ruff/architecture/
ai-readiness gates all clean.
